+++ delivery/infrastructure/src/main/resources/db/migration/V1_3__orders.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS Orders (
-    order_id BIGINT PRIMARY KEY UNIQUE,
-    customer_id INT,
+    order_id SERIAL PRIMARY KEY UNIQUE,
+    customer_id BIGINT,
     order_date TIMESTAMP,
     total_price DECIMAL(10,2),
     address VARCHAR(100),
